requirements.txt
django>=5.2,<6.0
djangorestframework>=3.15,<4.0
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
celery>=5.4,<6.0
django-celery-results>=2.5
redis>=5.0
psycopg2-binary>=2.9
django-storages[boto3]>=1.14
twilio>=9.0
django-ratelimit>=4.1
sentry-sdk>=2.0
gunicorn>=21.0
boto3>=1.34
